requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
schedule>=1.2.0
requests>=2.31.0
Pillow>=10.0.0
python-telegram-bot>=20.0
